chore: refactor parse function
